spear_rover: cleaned up C++ mapper
